add redirects to catalogs and move some firebase code to database/catalog.js
